agents-postgresql/requirements.txt
flask==3.0.0
openai==1.3.0
httpx==0.27.0
httpcore==1.0.4
pyodbc==5.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
azure-cosmos==4.5.0
psycopg2-binary==2.9.9